+++ FaunaDB.Client/Types/Field.cs
@@ -1,51 +1,18 @@
 ﻿using System.Collections.Generic;
 using System;
 
-using static FaunaDB.Types.Result;
-
 namespace FaunaDB.Types
 {
     /// <summary>
     /// A field extractor for a FaunaDB <see cref="Value"/>
     /// <para>
-    /// See <see cref="Value"/> and <see cref="Codec"/>
+    /// See <see cref="Value"/>
     /// </para>
     /// </summary>
     public sealed class Field<T>
     {
-        static Func<Value, IResult<IReadOnlyList<V>>> ToCollection<V>(Path path, Field<V> field)
-        {
-            return input => input.To(Codec.ARRAY).FlatMap(ToList(path, field));
-        }
-
-        static Func<IReadOnlyList<Value>, IResult<IReadOnlyList<V>>> ToList<V>(Path path, Field<V> field)
-        {
-            return values =>
-            {
-                var success = new List<V>();
-                var failures = new List<string>();
-
-                for (int i = 0; i < values.Count; i++)
-                {
-                    IResult<V> result = field.Get(values[i]);
-
-                    result.Match(
-                        Success: x => success.Add(x),
-                        Failure: reason => {
-                            Path subPath = path.SubPath(Path.From(i)).SubPath(field.path);
-                            failures.Add($"\"{subPath}\" {reason}");
-                        });
-                }
-
-                if (failures.Count > 0)
-                    return Fail<IReadOnlyList<V>>($"Failed to collect values: {string.Join(", ", failures)}");
-
-                return Success<IReadOnlyList<V>>(success);
-            };
-        }
-
-        Path path;
-        Func<Value, IResult<T>> codec;
+        internal readonly Path path;
+        internal readonly Func<Value, IResult<T>> codec;
 
         internal Field(Path path, Func<Value, IResult<T>> codec)
         {
@@ -62,12 +29,11 @@ namespace FaunaDB.Types
             new Field<U>(path.SubPath(other.path), other.codec);
 
         /// <summary>
-        /// Creates a field extractor that coerces its value using the codec passed
+        /// Creates a field extractor that coerces its value using the type specified
         /// </summary>
-        /// <param name="codec">codec to be used to coerce the field's value</param>
-        /// <returns>a new field that coerces its value using the codec passed</returns>
-        public Field<U> To<U>(Func<Value, IResult<U>> codec) =>
-            new Field<U>(path, codec);
+        /// <returns>a new field that coerces its value using the type specified</returns>
+        public Field<U> To<U>() =>
+            new Field<U>(path, Field.Decode<U>);
 
         /// <summary>
         /// Creates a field extractor that collects each inner value of an array using the nested field passed,
@@ -76,7 +42,7 @@ namespace FaunaDB.Types
         /// <param name="field">field to be extracted from each array's element</param>
         /// <returns>a new field that collects each inner value using the field passed</returns>
         public Field<IReadOnlyList<U>> Collect<U>(Field<U> field) =>
-            new Field<IReadOnlyList<U>>(path, ToCollection(path, field));
+            new Field<IReadOnlyList<U>>(path, Field.ToCollection(path, field));
 
         internal IResult<T> Get(Value root) =>
             path.Get(root).FlatMap(codec);
@@ -97,13 +63,54 @@ namespace FaunaDB.Types
     /// <summary>
     /// A field extractor for a FaunaDB <see cref="Value"/>
     /// <para>
-    /// See <see cref="Value"/> and <see cref="Codec"/>
+    /// See <see cref="Value"/>
     /// </para>
     /// </summary>
     public static class Field
     {
+        internal static Func<Value, IResult<IReadOnlyList<V>>> ToCollection<V>(Path path, Field<V> field)
+        {
+            return input => input.To<Value[]>().FlatMap(values =>
+            {
+                var success = new List<V>();
+                var failures = new List<string>();
+
+                for (int i = 0; i < values.Length; i++)
+                {
+                    IResult<V> result = field.Get(values[i]);
+
+                    result.Match(
+                        Success: x => success.Add(x),
+                        Failure: reason => {
+                            Path subPath = path.SubPath(Path.From(i)).SubPath(field.path);
+                            failures.Add($"\"{subPath}\" {reason}");
+                        });
+                }
+
+                if (failures.Count > 0)
+                    return Result.Fail<IReadOnlyList<V>>($"Failed to collect values: {string.Join(", ", failures)}");
+
+                return Result.Success<IReadOnlyList<V>>(success);
+            });
+        }
+
+        internal static IResult<T> Decode<T>(Value value)
+        {
+            try
+            {
+                if (value == NullV.Instance)
+                    return Result.Fail<T>("Value is null");
+
+                return Result.Success(Decoder.Decode<T>(value));
+            }
+            catch (Exception ex)
+            {
+                return Result.Fail<T>(ex.Message);
+            }
+        }
+
         internal static readonly Field<Value> Root =
-            new Field<Value>(Path.Empty, Codec.VALUE);
+            new Field<Value>(Path.Empty, Result.Success);
 
         /// <summary>
         /// Creates a field that extracts its value from a object path, assuming the value
@@ -112,7 +119,7 @@ namespace FaunaDB.Types
         /// <param name="keys">path to the field</param>
         /// <returns>the field extractor</returns>
         public static Field<Value> At(params string[] keys) =>
-            new Field<Value>(Path.From(keys), Codec.VALUE);
+            new Field<Value>(Path.From(keys), Result.Success);
 
         /// <summary>
         /// Creates a field that extracts its value from a array index, assuming the value
@@ -121,14 +128,13 @@ namespace FaunaDB.Types
         /// <param name="indexes">indexes path to the value</param>
         /// <returns>the field extractor</returns>
         public static Field<Value> At(params int[] indexes) =>
-            new Field<Value>(Path.From(indexes), Codec.VALUE);
+            new Field<Value>(Path.From(indexes), Result.Success);
 
         /// <summary>
-        /// Creates a field that coerces its value using the codec passed
+        /// Creates a field that coerces its value using the type specified.
         /// </summary>
-        /// <param name="codec">codec used to coerce the field's value</param>
         /// <returns>the field extractor</returns>
-        public static Field<T> To<T>(Func<Value, IResult<T>> codec) =>
-            new Field<T>(Path.Empty, codec);
+        public static Field<T> To<T>() =>
+            new Field<T>(Path.Empty, Decode<T>);
     }
 }